Tester
Slettet tomme Unit Test klasser og opprettet UITest av FeedbackClass
med recording i XCode.
