--- BotLer - Prototype 1Tests/FeedbackClassTest.swift
@@ -1,28 +1,0 @@
-//
-//  FeedbackClassTest.swift
-//  BotLer - Prototype 1
-//
-//  Created by Martha Honganvik Andersen on 08.03.2017.
-//  Copyright © 2017 Sondre Brekke. All rights reserved.
-//
-
-import XCTest
-
-class FeedbackClassTest: XCTestCase {
-    
-    override func setUp() {
-        super.setUp()
-        // Put setup code here. This method is called before the invocation of each test method in the class.
-    }
-    
-    override func tearDown() {
-        // Put teardown code here. This method is called after the invocation of each test method in the class.
-        super.tearDown()
-    }
-    
-    func testExample() {
-        XCTAssert(true)
-    }
-    
-    
-}
